requirements.txt
aiohttp>=3.9.0
groq>=0.11.0
python-dotenv>=1.0.1
edge-tts>=6.1.15
